Add files via upload
There are 5 requirements to have a romantic partner; singleness, maturity, intelligence, trust, love
All these requirements can take any numerical value between 1 and 10 except for singleness which can only be 1
Write a function, love that requires a user to input numerical values of singleness, maturity, intelligence, trust, love scores. Mind you that all requirements must have at least a numerical value of 1 but to be eligible for a romantic relationship, a person needs a total score of 21
Your function outputs, “Congrats! You are closer to your desired romantic relationship” when a person has a score of at least 21
Requirements:
Your code must require the user to input their scores as numbers only and return, “Kindly input a number between 1 and 10" when a type, number is not entered by the user
If a user inputs a number > or < 1 for singleness, your code should return, “Kindly input the number 1 here”
Your function should contain at least one type
There are at least 6 essentials here that you will be graded on
of control flow, that is, at least one type of loop and at least one conditional
Finally, it is important to note that your function does not take in any parameters